cloud/digitalocean: fix typo and move deployer functions
